feat: model de users implementado e service login criado

+++ backend/src/database/migrations/20230422151542-products.js
@@ -14,7 +14,7 @@ module.exports = {
         field: 'title'
       },
       price: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL(7, 2),
         allowNull: false,
         field: 'price',
       },
